Commit con los siguientes cambios: - Paginas que funcionan online al estar sin conexion o en modo offline   no permite hacer peticiones.

+++ src/containers/ListaRecibos/ListaReciboPendiente.js
@@ -45,36 +45,47 @@ export const ListaReciboPendiente = (props) => {
     }
 
     const cargarRecibos = async () => {
-        fetch(urlApi + "/api/Recibo/Pendiente", {
-            headers: {
-                'Authorization':
-                    'Bearer ' + localStorage.getItem('token')
+        let isOnline = await verificarConexion();
+        if (!isOnline) {
+            Swal.fire({
+                title: "Sin internet",
+                text: "Necesita internet para poder visualizar esta pagina.",
+                type: "warning",
+                confirmButtonText: 'Ok',
+            });
+            setIsLoaded(true);
+        } else {
+            fetch(urlApi + "/api/Recibo/Pendiente", {
+                headers: {
+                    'Authorization':
+                        'Bearer ' + localStorage.getItem('token')
 
-            }
-        })
-            .then(res => {
-                if (res.status === 401) {
-                    localStorage.setItem('token', '');
-                    window.location.reload();
-                }
-                if (res.status === 200) {
-                    res.json()
-                        .then(
-                            (result) => {
-
-                                setRecibos(result);
-                                setIsLoaded(true);
-                            },
-                            // Note: it's important to handle errors here
-                            // instead of a catch() block so that we don't swallow
-                            // exceptions from actual bugs in components.
-                            (error) => {
-                                setError(error);
-                                setIsLoaded(true);
-                            }
-                        )
                 }
             })
+                .then(res => {
+                    if (res.status === 401) {
+                        localStorage.setItem('token', '');
+                        window.location.reload();
+                    }
+                    if (res.status === 200) {
+                        res.json()
+                            .then(
+                                (result) => {
+
+                                    setRecibos(result);
+                                    setIsLoaded(true);
+                                },
+                                // Note: it's important to handle errors here
+                                // instead of a catch() block so that we don't swallow
+                                // exceptions from actual bugs in components.
+                                (error) => {
+                                    setError(error);
+                                    setIsLoaded(true);
+                                }
+                            )
+                    }
+                })
+        }
     }
 
 
@@ -123,7 +134,7 @@ export const ListaReciboPendiente = (props) => {
         if (recibo.anticipo) {
             ruta = `${urlApi}/api/Recibo/Anticipo/Pendiente/${recibo.recibo}`;
         }
-        if (isOnline) {
+        if (isOnline && localStorage.getItem("Conexion")==="Online") {
             try {
                 setLoading(true);
                 const request = await axios.post(ruta);
